requirements.txt
alembic==1.10.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.2.4
MarkupSafe==2.1.1
packaging==24.0
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.0.1
zipp==3.15.0
python-dotenv==1.0.0